+++ app/src/main/AndroidManifest.xml
@@ -12,16 +12,16 @@
         android:theme="@style/Theme.Dungeon"
         tools:targetApi="31">
         <activity
-            android:name=".MapActivity"
+            android:name=".view.MapActivity"
             android:exported="false" />
         <activity
-            android:name=".ItemActivity"
+            android:name=".view.ItemActivity"
             android:exported="false" />
         <activity
-            android:name=".EnemyActivity"
+            android:name=".view.EnemyActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
