+++ app/src/main/java/edu/utsa/cs3443/dungeon/model/Enemy.java
@@ -1,4 +1,4 @@
-package edu.utsa.cs3443.dungeon;
+package edu.utsa.cs3443.dungeon.model;
 
 /**
  */
@@ -6,8 +6,8 @@ public class Enemy extends Entity {
 
     /**
      * Defines the Entity class and relevant methods
-     * @param whatever idk
-     */
+     * @param _maxHP max hp
+     **/
     public Enemy(final String _name, final int _maxHP, final char _smallCharacter)
     {
         super(_name, 0, _maxHP, _smallCharacter);
