+++ app/src/main/java/edu/utsa/cs3443/dungeon/model/Player.java
@@ -1,4 +1,4 @@
-package edu.utsa.cs3443.dungeon;
+package edu.utsa.cs3443.dungeon.model;
 
 /**
  */
