+++ app/src/main/java/edu/utsa/cs3443/dungeon/view/MapActivity.java
@@ -1,8 +1,10 @@
-package edu.utsa.cs3443.dungeon;
+package edu.utsa.cs3443.dungeon.view;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import edu.utsa.cs3443.dungeon.R;
 
 /**
  */
